updates functionalities
